add + refactor test cases
